% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.r
\name{intsum}
\alias{intsum}
\title{Convert to int and sum}
\usage{
intsum(x, y, z)
}
\arguments{
\item{x}{int}

\item{y}{numeric}

\item{z}{char}
}
\description{
Convert to int and sum
}
\examples{
intsum( 3, 3.3, "abc")
}
